npis: Vectorized NPIs for R

An S3 vctrs class representing a National Provider Identifier, a unique 10-digit identification number assigned to healthcare providers in the United States, as mandated by the Health Insurance Portability and Accountability Act (HIPAA).

NPI Requirements

The NPI must meet the specifications detailed in the Final Rule for the Standard Unique Health Identifier for Health Care Providers (69 FR 3434), published 2004-01-23.

The NPI must consist of 9 numeric digits followed by 1 numeric check digit.

The NPI must not have embedded intelligence.

The NPI format and check digit calculation must be compatible with the card issuer identifier on a standard health identification card. As each card issuer identifier must be unique, the NPI must avoid collisions with other possible card issuer identifiers. NPIs will therefore initially be issued with a first digit of 1 or 2. These digits will be exclusive to the NPI.

Within the constraints of (4), NPI generation must employ a scattering algorithm that uses all possible numeric combinations beginning with a 1 or 2.

Each NPI generated must be unique, without requiring database access for verification.

NPI Check Digit

The NPI check digit is calculated using the Luhn algorithm for computing the modulus 10 "double-add-double" check digit, a recognized ISO standard (ISO/IEC 7812).

When an NPI is used as a card issuer identifier, it is prefixed by 80840, in which 80 indicates Health Applications and 840 indicates United States.

The check digit will always be calculated as if the prefix is present. This is accomplished by simply adding the constant 24 in the check digit calculation when the NPI is used without the prefix.

Source

PDF: Requirements for National Provider Identifier (NPI) and NPI Check Digit

See also

Author

Maintainer: Andrew Allen Bruce andrewallenbruce@gmail.com [copyright holder]